apps/requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.2.0
Django==5.2.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.5
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
